+++ frontend/src/main.js
@@ -3,6 +3,7 @@ import App from './App.vue'
 import { createI18n } from 'vue-i18n'
 import router from './router'
 import { registerSW } from 'virtual:pwa-register'
+import { createHead } from '@unhead/vue'
 
 registerSW({ immediate: true })
 const i18n = createI18n({
@@ -16,7 +17,9 @@ const i18n = createI18n({
         messages: {}
     }
 })
+const head = createHead()
 const app = createApp(App)
 app.use(i18n)
 app.use(router)
+app.use(head)
 app.mount('#app')